refactored review section into components and fixed issue with review text showing under review header on scroll
