+++ client/src/components/rating/components/ReviewRec.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+export default function ReviewRec({ rec }) {
+  let recommended;
+
+        if (rec) {
+          recommended = (
+            <div>
+              <i className="fa-solid fa-check"></i> I recommend this product.
+            </div>
+          );
+        }
+
+  return recommended;
+}
