+++ client/src/components/rating/helpers/isImage.js
@@ -1,0 +1,5 @@
+const isImage = (url) => {
+    return /\.(jpg|jpeg|png|webp|avif|gif|svg)$/.test(url);
+  };
+
+module.exports = isImage;
